improved PR
